Moved common functions and values to a utilities file in the integrationTests package
Got AWS and Azure tests working.
