+++ integrationTests/azureUpload_test.go
@@ -2,18 +2,14 @@ package integrationTests
 
 import (
 	"github.com/jedi91/mob-stor/azure"
-	"github.com/jedi91/mob-stor/distribute"
 	"github.com/jedi91/mob-stor/transmit"
 	"os"
 	"testing"
 	"time"
 )
 
-const boolTemplate = "Actual: %t | Expected: %t"
-const dateFormat = "01-02-2006 15:04:05"
-
 func TestAzureFileUpload(t *testing.T) {
-	transmitter := setupTransmitter()
+	transmitter := setupAzureTransmitter()
 	distributor := setupDistributor(
 		transmitter,
 	)
@@ -44,7 +40,7 @@ func TestAzureFileUpload(t *testing.T) {
 }
 
 func TestAzureFileUploadWithPath(t *testing.T) {
-	transmitter := setupTransmitter()
+	transmitter := setupAzureTransmitter()
 	distributor := setupDistributor(
 		transmitter,
 	)
@@ -74,7 +70,7 @@ func TestAzureFileUploadWithPath(t *testing.T) {
 	)
 }
 
-func setupTransmitter() transmit.Transmitter {
+func setupAzureTransmitter() transmit.Transmitter {
 	credsProvider := azure.BlobCredentialProvider{
 		AccountName: os.Getenv("AZURE_STORAGE_ACCOUNT"),
 		AccountKey:  os.Getenv("AZURE_STORAGE_ACCESS_KEY"),
@@ -87,43 +83,4 @@ func setupTransmitter() transmit.Transmitter {
 	return transmit.AzureBlobTransmitter{
 		ContainerUrlProvider: urlProvider,
 	}
-}
-
-func setupDistributor(
-	transmitter transmit.Transmitter,
-) distribute.Distributor {
-	transmitters := []transmit.Transmitter{
-		transmitter,
-	}
-
-	return distribute.Distributor{
-		Transmitters: transmitters,
-	}
-}
-
-func checkForSuccess(
-	results []distribute.Result,
-) bool {
-	success := true
-	for _, result := range results {
-		success = success && result.Success
-	}
-
-	return success
-}
-
-func checkExpectedBool(
-	t *testing.T,
-	actual bool,
-	expected bool,
-) {
-	if actual == expected {
-		return
-	}
-
-	t.Errorf(
-		boolTemplate,
-		actual,
-		expected,
-	)
 }
